fix type of DeltaX

+++ WebApi/Entities/Function.cs
@@ -3,7 +3,7 @@ namespace WebApi.Entities {
         public long Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public float DeltaX { get; set; }
+        public double DeltaX { get; set; }
         public bool IsNeedMaxLoadFactor { get; set; }
     }
 }
